:lipstick: #45 add next level button in score view
