+++ src/components/InterfaceScore/index.scss
@@ -46,7 +46,7 @@
     position: relative;
   }
 
-  &__min-area {
+  &__area {
     width: 100%;
     display: flex;
     flex-direction: column;
@@ -57,7 +57,7 @@
   }
 
   &__panel {
-    bottom: 37%;
+    bottom: 36%;
     position: absolute;
   }
 
@@ -70,6 +70,11 @@
     text-align: center;
     text-transform: uppercase;
     @include mikado-black;
+  }
+
+  &__next {
+    position: absolute;
+    bottom: 31%;
   }
 
   &__animation {
