+++ src/components/InterfaceScore/index.tsx
@@ -1,12 +1,15 @@
-import React, { FC, useEffect, useState } from "react";
+import React, { FC, useEffect, useState, useCallback } from "react";
+import { useTranslation } from "react-i18next";
 
 // components
+import InterfaceButton, { Colors } from "../../components/InterfaceButton";
 import InterfaceScorePanel from "../../components/InterfaceScorePanel";
 import Area from "../../components/Area";
 
 // store
 import { useTypedSelector } from "../../redux/store";
-import { selectors } from "../../redux";
+import { useDispatch, useSelector } from "react-redux";
+import { actions, selectors } from "../../redux";
 
 // styles
 import "./index.scss";
@@ -21,6 +24,15 @@ const InterfaceScore: FC<Props> = ({ isActive }) => {
   const isSuccess = useTypedSelector(selectors.round.selectIsSuccess);
   const areaMinHeight = useTypedSelector(selectors.area.selectMinHeight);
   const isCreator = useTypedSelector(selectors.room.selectIsCreator);
+
+  const dispatch = useDispatch();
+  const { t } = useTranslation();
+
+  // handlers
+
+  const handleOnNextRoundClick = useCallback(() => {
+    dispatch(actions.webSocket.emit.transition.ended());
+  }, [dispatch]);
 
   // state
 
@@ -44,7 +56,7 @@ const InterfaceScore: FC<Props> = ({ isActive }) => {
           }}
         ></div>
         <div
-          className="score__min-area"
+          className="score__area"
           style={{
             height: areaMinHeight
           }}
@@ -67,14 +79,20 @@ const InterfaceScore: FC<Props> = ({ isActive }) => {
         <Area height="min">
           <div className="score__bush"></div>
         </Area>
-        <div
-          className="score__min-area"
-          style={{
-            height: areaMinHeight
-          }}
-        >
-          <div className="score__animation">
-            {isCreator && (
+        {isCreator && (
+          <div
+            className="score__area"
+            style={{
+              height: areaMinHeight
+            }}
+          >
+            <InterfaceButton
+              onClick={handleOnNextRoundClick}
+              color={Colors.blue}
+              text={t("score.buttons.next")}
+              classNames="score__next"
+            />
+            <div className="score__animation">
               <img
                 className="score__gif"
                 src={require(`../../assets/illustrations/score/gifs/${
@@ -82,9 +100,9 @@ const InterfaceScore: FC<Props> = ({ isActive }) => {
                 }.gif`)}
                 alt="Gif"
               />
-            )}
+            </div>
           </div>
-        </div>
+        )}
       </div>
     </div>
   );
